[MIG] multi_store modules: Migration to 17.0
