+++ account_multi_store/models/res_store.py
@@ -23,8 +23,6 @@ class ResStore(models.Model):
     only_allow_reonciliaton_of_this_store = fields.Boolean(
         help='If enable, debt reconciliation will be only doable on items of this store')
 
-    dont_compute_withholdings = fields.Boolean()
-
     @api.depends('journal_ids')
     def _compute_journals_count(self):
         for rec in self:
@@ -37,25 +35,3 @@ class ResStore(models.Model):
                 rec.currency_exchange_journal_id.store_id = rec.id
             elif rec.currency_exchange_journal_id.store_id != rec.id:
                 raise UserError('El diario de diferencia de cambio debe ser de esta misma sucursal')
-
-    @api.model
-    def _init_payment_group_multi_store_demo(self):
-
-        # por ahora lo estamos haciendo solo para main company
-        main_company = self.env.ref('base.main_company')
-
-        store_a = self.env.ref('account_multi_store.unidad_a_store')
-        # si ya configuramos diarios, seguimos adelante
-        if store_a.journal_ids:
-            return
-
-        journals = self.env[('account.journal')].search([('company_id', '=', main_company.id)])
-        journals.store_id = store_a.id
-        
-        # duplicamos todos los diarios de liquidez, venta, compra y el de diferencia de cambio
-        store_b = self.env.ref('account_multi_store.unidad_b_store')
-        for journal in journals.filtered(lambda x: x.type in ['sale', 'purchase', 'bank', 'cash'] or x.code == 'EXCH'):
-            journal_b = journal.copy(default={'store_id': store_b.id})
-            journal_b.name = '%s - B' % (journal.name)
-            journal.name += ' - A'
-
