+++ account_payment_pro_multi_store/demo/res_store_demo.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="account_multi_store.unidad_b_store" model="res.store">
+        <field name="dont_compute_withholdings" eval="True"/>
+    </record>
+
+    <function model="res.store" name="_init_payment_group_multi_store_demo"/>
+    
+</odoo>
