+++ account_payment_pro_multi_store/models/__init__.py
@@ -1,2 +1,4 @@
 from . import account_payment
 from . import account_move
+from . import account_move_line
+from . import res_store
